moved admin and post outside dashboard(navigation)
