+++ src/app/admin/post-list/post-list.component.scss
