+++ src/app/app-routing.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { AddPostComponent } from './dashboard/admin/add-post/add-post.component';
-import { AdminComponent } from './dashboard/admin/admin.component';
-import { EditPostComponent } from './dashboard/admin/edit-post/edit-post.component';
-import { PostListComponent } from './dashboard/admin/post-list/post-list.component';
-import { PostComponent } from './dashboard/post/post.component';
+import { AddPostComponent } from './admin/add-post/add-post.component';
+import { AdminComponent } from './admin/admin.component';
+import { EditPostComponent } from './admin/edit-post/edit-post.component';
+import { PostListComponent } from './admin/post-list/post-list.component';
+import { PostComponent } from './post/post.component';
 import { HomeLayoutComponent } from './home-layout/home-layout.component';
 import { LoginLayoutComponent } from './login-layout/login-layout.component';
 import { LoginComponent } from './login/login.component';
