+++ src/app/app.module.ts
@@ -5,17 +5,17 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { DashboardComponent } from './dashboard/dashboard.component';
-import { PostComponent } from './dashboard/post/post.component';
-import { AdminComponent } from './dashboard/admin/admin.component';
+import { PostComponent } from './post/post.component';
+import { AdminComponent } from './admin/admin.component';
 import { TruncatePipe } from './shared/pipes/truncate.pipe';
 import { FooterComponent } from './footer/footer.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MaterialModule } from './shared/material/material.module';
-import { AddPostComponent } from './dashboard/admin/add-post/add-post.component';
+import { AddPostComponent } from './admin/add-post/add-post.component';
 import { HttpClientModule } from '@angular/common/http';
-import { EditPostComponent } from './dashboard/admin/edit-post/edit-post.component';
+import { EditPostComponent } from './admin/edit-post/edit-post.component';
 import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
-import { PostListComponent } from './dashboard/admin/post-list/post-list.component';
+import { PostListComponent } from './admin/post-list/post-list.component';
 import { DatePipe } from '@angular/common';
 import { LoginComponent } from './login/login.component';
 import { LoginLayoutComponent } from './login-layout/login-layout.component';
