+++ src/app/post/post.component.scss
