+++ src/app/post/post.component.ts
